backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
SQLAlchemy==2.0.23
PyMySQL==1.1.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
textblob==0.17.1
nltk==3.8.1
matplotlib==3.8.2
seaborn==0.13.0
joblib==1.3.2
python-dotenv==1.0.0
itsdangerous==2.1.2
